Create Data_cleaning.sql
As part of theforage.com's GE Aviation virtual internship, you are asked to combine multiple datasets into one usable table in excel and modify one of the columns to put the temperature measurements into Rankine. This is easy enough to do in excel, but could it be done in SQL? The code below demonstrates one way this could be accomplished using small amounts of data from the database for ease of use.